+++ src/main/java/com/chatter/Chatly/domain/article/ArticleService.java
@@ -9,9 +9,9 @@ import com.chatter.Chatly.domain.channel.Channel;
 import com.chatter.Chatly.domain.channel.ChannelRepository;
 import com.chatter.Chatly.domain.member.Member;
 import com.chatter.Chatly.domain.member.MemberRepository;
-import com.chatter.Chatly.dto.ArticleDto;
-import com.chatter.Chatly.dto.ArticleRequestDto;
-import com.chatter.Chatly.dto.TargetsDto;
+import com.chatter.Chatly.domain.article.dto.ArticleDto;
+import com.chatter.Chatly.domain.article.dto.ArticleRequestDto;
+import com.chatter.Chatly.domain.common.dto.TargetsDto;
 import com.chatter.Chatly.exception.CommonErrorCode;
 import com.chatter.Chatly.exception.HttpException;
 import com.chatter.Chatly.util.MemberContext;
