+++ src/main/java/com/chatter/Chatly/domain/channel/ChannelService.java
@@ -9,8 +9,8 @@ import com.chatter.Chatly.domain.channelmember.ChannelMemberService;
 import com.chatter.Chatly.domain.chatroom.ChatRoom;
 import com.chatter.Chatly.domain.chatroom.ChatRoomRepository;
 import com.chatter.Chatly.domain.common.Role;
-import com.chatter.Chatly.dto.ChannelDto;
-import com.chatter.Chatly.dto.ChannelRequestDto;
+import com.chatter.Chatly.domain.channel.dto.ChannelDto;
+import com.chatter.Chatly.domain.channel.dto.ChannelRequestDto;
 import com.chatter.Chatly.exception.CommonErrorCode;
 import com.chatter.Chatly.exception.HttpException;
 import com.chatter.Chatly.util.MemberContext;
