+++ src/main/java/com/chatter/Chatly/domain/channelmember/dto/RoleRequestDto.java
@@ -1,4 +1,4 @@
-package com.chatter.Chatly.dto;
+package com.chatter.Chatly.domain.channelmember.dto;
 
 import java.util.List;
