+++ src/main/java/com/chatter/Chatly/domain/comment/CommentService.java
@@ -13,8 +13,8 @@ import com.chatter.Chatly.domain.article.Article;
 import com.chatter.Chatly.domain.article.ArticleRepository;
 import com.chatter.Chatly.domain.member.Member;
 import com.chatter.Chatly.domain.member.MemberRepository;
-import com.chatter.Chatly.dto.CommentDto;
-import com.chatter.Chatly.dto.CommentRequestDto;
+import com.chatter.Chatly.domain.comment.dto.CommentDto;
+import com.chatter.Chatly.domain.comment.dto.CommentRequestDto;
 import com.chatter.Chatly.exception.CommonErrorCode;
 import com.chatter.Chatly.exception.HttpException;
 import com.chatter.Chatly.util.MemberContext;
