+++ src/main/java/com/chatter/Chatly/domain/comment/dto/CommentDto.java
@@ -1,4 +1,4 @@
-package com.chatter.Chatly.dto;
+package com.chatter.Chatly.domain.comment.dto;
 
 import java.time.LocalDateTime;
