+++ src/main/java/com/chatter/Chatly/domain/like/LikeController.java
@@ -1,19 +1,10 @@
 package com.chatter.Chatly.domain.like;
 
-import com.chatter.Chatly.annotation.CheckAccessPossession;
-import com.chatter.Chatly.annotation.RequireOwnership;
-import com.chatter.Chatly.annotation.RequirePrivilege;
-import com.chatter.Chatly.domain.chatroom.ChatRoomService;
 import com.chatter.Chatly.domain.like.dto.LikeDto;
 import com.chatter.Chatly.domain.like.dto.LikeRequestDto;
-import com.chatter.Chatly.dto.ChatRoomDto;
-import com.chatter.Chatly.dto.ChatRoomRequestDto;
 import lombok.RequiredArgsConstructor;
-import org.apache.coyote.Response;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/like/{cid}")
