+++ src/main/java/com/chatter/Chatly/global/auth/AuthController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.chatter.Chatly.dto.MemberRequestDto;
+import com.chatter.Chatly.domain.member.dto.MemberRequestDto;
 
 
 @RestController
