+++ src/test/java/com/chatter/Chatly/domain/article/ArticleControllerTest.java
@@ -29,9 +29,9 @@ import com.chatter.Chatly.domain.channelmember.ChannelMember;
 import com.chatter.Chatly.domain.channelmember.ChannelMemberService;
 import com.chatter.Chatly.domain.common.Role;
 import com.chatter.Chatly.domain.member.Member;
-import com.chatter.Chatly.dto.ArticleDto;
-import com.chatter.Chatly.dto.ArticleRequestDto;
-import com.chatter.Chatly.dto.TargetsDto;
+import com.chatter.Chatly.domain.article.dto.ArticleDto;
+import com.chatter.Chatly.domain.article.dto.ArticleRequestDto;
+import com.chatter.Chatly.domain.common.dto.TargetsDto;
 import com.chatter.Chatly.testUtil.TestEntitySetter;
 import com.chatter.Chatly.util.JwtUtil;
 import com.fasterxml.jackson.databind.ObjectMapper;
