+++ src/test/java/com/chatter/Chatly/domain/channelmember/ChannelMemberControllerTest.java
@@ -17,7 +17,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
@@ -35,9 +34,9 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.chatter.Chatly.domain.channel.Channel;
 import com.chatter.Chatly.domain.common.Role;
 import com.chatter.Chatly.domain.member.Member;
-import com.chatter.Chatly.dto.ChannelMemberDto;
-import com.chatter.Chatly.dto.ChannelMemberRequestDto;
-import com.chatter.Chatly.dto.RoleRequestDto;
+import com.chatter.Chatly.domain.channelmember.dto.ChannelMemberDto;
+import com.chatter.Chatly.domain.channelmember.dto.ChannelMemberRequestDto;
+import com.chatter.Chatly.domain.channelmember.dto.RoleRequestDto;
 import com.chatter.Chatly.exception.CommonErrorCode;
 import com.chatter.Chatly.exception.HttpException;
 import com.chatter.Chatly.testUtil.TestEntitySetter;
